+++ parser/test_pas.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 run() {
-	cc -lreadline  ./*.c ./main_pars/main.c ../lexer/*.c ../libft/libft.a 
+	cc -lreadline  ./*.c ./main_pars/main.c ../lexer/*.c ../libft/libft.a ../utils/fprintf/fprintf.a
 	input="$1"
 	# leak_check=(valgrind -q --leak-check=full)
 	#if you want to check leaking, add this command "${leak_check[0]}" to head of ↓
